feat/19-minor-integrations
