+++ src/App.jsx
@@ -10,7 +10,6 @@ import LeavesPage from "./pages/LeavesPage";
 import HolidaysPage from "./pages/HolidaysPage";
 import ShiftsPage from "./pages/ShiftsPage";
 import NotFoundPage from "./pages/NotFoundPage";
-import UserTable from "./components/UserTable";
 
 function App() {
   const router = createBrowserRouter([
@@ -51,7 +50,7 @@ function App() {
       ),
     },
     {
-      path: "/shifts",
+      path: "/shift-policies",
       element: (
         <Protected>
           <ShiftsPage />
