+++ src/components/Header.jsx
@@ -1,5 +1,5 @@
 import Button from "./Button";
-import logo from "../assets/flipopay_logo.png";
+import logo from "../assets/flipopay-logo.png";
 import { signOut, getAuth } from "firebase/auth";
 
 const Header = () => {
@@ -16,11 +16,7 @@ const Header = () => {
   return (
     <header className="w-full drop-shadow-xl bg-sky-200 flex  mb-4  justify-between items-center h-16 shadow-md">
       <div className="flex items-center">
-        <img
-          src={logo} // Replace with your logo URL
-          alt="Logo"
-          className="h-16 py-4 px-2 ml-2"
-        />
+        <img src={logo} alt="Logo" className="h-16 py-4 px-2 ml-2" />
       </div>
       <h1 className="left-1/2 transform -translate-x-1/2 text-xl font-bold">
         Attendance Management System
