+++ src/pages/AdminPage.jsx
@@ -1,8 +1,9 @@
 import StyledDatePicker from "../components/StyledDatePicker";
 import SearchBar from "../components/SearchBar";
-import Recordstable from "../components/Recordstable";
+import Recordstable from "../components/table/Recordstable";
 import CreateEmployeeModal from "../components/modals/CreateEmployeeModal";
 import PageOutline from "../components/PageOutline";
+import { useState } from "react";
 
 export default function AdminPage() {
   const currentDate = new Date();
