+++ src/pages/HolidaysPage.jsx
@@ -1,6 +1,6 @@
 import StyledTH from "../components/StyledTH";
 import StyledTD from "../components/StyledTD.jsx";
-import Table from "../components/Table.jsx";
+import Table from "../components/table/Table.jsx";
 import PageOutline from "../components/PageOutline.jsx";
 import AddHolidayModal from "../components/modals/AddHolidayModal.jsx";
 import useFetchCollection from "../hooks/UseFetchCollection.jsx";
